examples/eidosui-components/pyproject.toml
[project]
name = "eidosui-components"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "air>=0.10.0",
    "eidosui>=0.2.0",
]
